Rearrange files and make opt public
