+++ ribs/emitters/opt/_gradient_opt_base.py
@@ -1,0 +1,5 @@
+"""Provides GradientOptBase."""
+
+
+class GradientOptBase:
+    """Base class for gradient-based optimizers."""
